refactor(fundraiser): fix delete public procedure and error
